refactor(backend): split diary and intent services
