+++ docs/plans/漫画功能更新记录_20260313-14.md
@@ -153,3 +153,9 @@
 | 生成触发 | 单独按钮 | 与 AI 文字并发生成 |
 | 日期切换 | 手动刷新 | 自动加载已有结果 |
 | 架构 | `lifetrace/` 旧目录 | `server/` 新架构 + 热重载 cron |
+
+### 当前后端结构补充
+
+- `server/services/diary_illustration_service.py` 现在主要保留服务编排、状态管理和公开入口。
+- Provider 选择、供应商调用和调度辅助逻辑已拆到 `server/services/diary_illustration_provider.py`。
+- Prompt 解析和图片生成辅助逻辑已拆到 `server/services/diary_illustration_generation.py`。
